# docs/swagger.yaml
basePath: /api/v1
definitions:
  request.EditProfileRequest:
    properties:
      avatar_id:
        type: integer
      gender:
        type: string
      name:
        type: string
    type: object
  request.LoginRequest:
    properties:
      email:
        example: satria@gmail.com
        type: string
      password:
        example: satria123
        type: string
    required:
    - email
    - password
    type: object
  request.OnboardingRequest:
    properties:
      avatar_id:
        example: 1
        maximum: 4
        minimum: 1
        type: integer
      gender:
        enum:
        - male
        - female
        example: male
        type: string
    required:
    - avatar_id
    - gender
    type: object
  request.RegisterRequest:
    properties:
      email:
        example: satria@gmail.com
        type: string
      name:
        example: Gantang Satria
        type: string
      password:
        example: satria123
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
...
  response.AuthResponse:
    properties:
      access_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      user:
        $ref: '#/definitions/response.UserData'
    type: object
...
  response.Meta:
    properties:
      message:
        example: operasi berhasil
        type: string
      success:
        example: true
        type: boolean
    type: object
...
  response.Response:
    properties:
      data: {}
      meta:
        $ref: '#/definitions/response.Meta'
    type: object
...
  response.UserData:
    properties:
      avatar_id:
        example: 1
        type: integer
      gender:
        example: male
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      name:
        example: Gantang Satria
        type: string
    type: object
  response.UserResponse:
    properties:
      avatar_id:
        example: 1
        type: integer
      current_streak:
        example: 3
        type: integer
      gender:
        example: male
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      longest_streak:
        example: 7
        type: integer
      name:
        example: Gantang Satria
        type: string
      total_xp:
        type: integer
    type: object
host: empathify-test.fly.dev
info:
  contact:
    email: empathify@gmail.com
    name: Empathify Team
  description: API untuk platform terapi dan pelatihan keterampilan sosial
  termsOfService: http://swagger.io/terms/
  title: Empathify API
  version: "1.0"
paths:
  /auth/google:
    get:
      description: Redirect user ke halaman login Google OAuth
      responses:
        "302":
          description: Found
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Login dengan Google
      tags:
      - Auth
  /auth/google/callback:
    get:
      description: Callback dari Google setelah user berhasil login. Otomatis register
        jika akun belum ada.
      parameters:
      - description: Authorization code dari Google
        in: query
        name: code
        required: true
        type: string
      - description: State untuk CSRF protection
        in: query
        name: state
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.AuthResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Google OAuth callback
      tags:
      - Auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login menggunakan email dan password yang sudah terdaftar
      parameters:
      - description: Login payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.AuthResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Login dengan email
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Daftarkan akun baru menggunakan email dan password. Gender dan
        avatar diisi saat onboarding.
      parameters:
      - description: Register payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.AuthResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Register dengan email
      tags:
      - Auth
...
  /user/onboarding:
    patch:
      consumes:
      - application/json
      description: Isi gender dan avatar setelah register email atau login Google
        pertama kali
      parameters:
      - description: Onboarding payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.OnboardingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Onboarding user
      tags:
      - User
  /user/profile:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.UserResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Ambil profil user
      tags:
      - User
    patch:
      consumes:
      - application/json
      description: Update nama, gender, atau avatar. Semua field opsional.
      parameters:
      - description: Data profil yang ingin diubah
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.EditProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Edit profil user
      tags:
      - User
securityDefinitions:
  BearerAuth:
    description: 'Masukkan token dengan format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"