# pkg/dto/request/user_request.go
package request

type OnboardingRequest struct {
	Gender   string `json:"gender" validate:"required,oneof=male female" example:"male"`
	AvatarID int16  `json:"avatar_id" validate:"required,min=1,max=4" example:"1"`
}

type EditProfileRequest struct {
	Name     *string `json:"name"`
	Gender   *string `json:"gender"`
	AvatarID *int16  `json:"avatar_id"`
}